requirements.txt
asgiref==3.5.1
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.4.2
Pillow==9.1.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1